+++ .vscode/settings.json
@@ -1,11 +1,15 @@
 {
     "code-runner.executorMap": {
-        "cpp": "cd $dir && clang++ -std=c++14 -Wall -Wno-missing-braces -Wno-unused-variable -Wno-unused-parameter -Wextra -pthread $fileName -o  ../bin/$fileNameWithoutExt && ../bin/$fileNameWithoutExt"
+        "cpp": "cd $dir && clang++ -std=c++14 -Wall -Wno-missing-braces -Wno-unused-variable -Wno-unused-parameter -Wno-unused-local-typedefs -pthread $fileName crtbp.cpp pcrtbp.cpp -o ../bin/$fileNameWithoutExt && ../bin/$fileNameWithoutExt"
     },
     "files.associations": {
+        "*.backup": "plaintext",
         "*.dat": "fortran-modern",
+        "*.fragmentshader": "glsl",
         "*.in": "fortran-modern",
         "*.m": "matlab",
+        "*.man": "plaintext",
+        "*.vertexshader": "glsl",
         "*.vtk": "fortran-modern",
         "*.ipp": "cpp",
         "cmath": "cpp",
@@ -40,6 +44,7 @@
         "limits": "cpp",
         "deque": "cpp",
         "map": "cpp",
-        "ios": "cpp"
+        "ios": "cpp",
+        "sstream": "cpp"
     }
 }
