--- src/tempCodeRunnerFile.cpp
@@ -1,2 +1,0 @@
-    func.read_newcomb();
-    func.simple_test();
